fix: add setTimeout to delay navigate function execution (line 184 and 190)
