+++ src/index.js
@@ -7,16 +7,14 @@ import reportWebVitals from "./reportWebVitals";
 //import Context
 import { AuthContextProvider } from "context/AuthContext";
 import { ModalContextProvider } from "context/ModalContext";
-import { NavigationContextProvider } from "context/NavigationContext";
+
 
 const root = ReactDOM.createRoot(document.getElementById("root"));
 root.render(
   <React.StrictMode>
     <AuthContextProvider>
-      <ModalContextProvider>
-        <NavigationContextProvider>
+      <ModalContextProvider> 
           <App />
-        </NavigationContextProvider>
       </ModalContextProvider>
     </AuthContextProvider>
   </React.StrictMode>
