+++ src/pages/UserOtherFollowPage.jsx
@@ -7,6 +7,8 @@ import LeftNavBar from "components/LeftNavBar/LeftNavBar";
 import MiddleInfo from "components/MiddleInfo/MiddleInfo";
 import RightBanner from "components/RightBanner/RightBanner";
 import UserOtherFollowPageInfo from "components/UserOtherFollowPageInfo/UserOtherFollowPageInfo";
+import { NavigationContextProvider } from "context/NavigationContext";
+// 因為NavigationContextProvider裡面的navigate是react-router-dom套件的函式，只能在router裡使用
 const UserOtherFollowPage = () => {
   return (
     <MainContainer>
@@ -15,7 +17,9 @@ const UserOtherFollowPage = () => {
       </LeftContainer>
       <MiddleContainer>
         <MiddleInfo>
-          <UserOtherFollowPageInfo />
+          <NavigationContextProvider>
+            <UserOtherFollowPageInfo />
+          </NavigationContextProvider>
         </MiddleInfo>
       </MiddleContainer>
       <RightContainer>
