+++ src/pages/UserSelfPage.jsx
@@ -7,7 +7,8 @@ import LeftNavBar from "components/LeftNavBar/LeftNavBar";
 import MiddleInfo from "components/MiddleInfo/MiddleInfo";
 import RightBanner from "components/RightBanner/RightBanner";
 import UserSelfPageInfo from "components/UserSelfPageInfo/UserSelfPageInfo";
-
+import { NavigationContextProvider } from "context/NavigationContext";
+// 因為NavigationContextProvider裡面的navigate是react-router-dom套件的函式，只能在router裡使用
 const UserSelfPage = () => {
   return (
     <MainContainer>
@@ -16,7 +17,9 @@ const UserSelfPage = () => {
       </LeftContainer>
       <MiddleContainer>
         <MiddleInfo>
-          <UserSelfPageInfo />
+          <NavigationContextProvider>
+            <UserSelfPageInfo />
+          </NavigationContextProvider>
         </MiddleInfo>
       </MiddleContainer>
       <RightContainer>
